refact: removed ResponseEntity from Controller and some DTO variable adjustment

+++ src/main/java/io/GuiWEspinola/poc1/entities/dto/response/CustomerResponseDTO.java
@@ -16,5 +16,5 @@ public class CustomerResponseDTO {
 
     private DocumentType documentType;
 
-    private Integer mobileNumber;
+    private Long mobileNumber;
 }
